requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
alembic==1.16.3
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
black==24.3.0
certifi==2025.7.9
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distlib==0.3.9
docopt==0.6.2
docstring_parser_fork==0.0.12
einops==0.8.1
executing==2.2.0
filelock==3.18.0
flake8==7.0.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
graphviz==0.21
greenlet==3.2.3
huggingface-hub==0.33.2
humanfriendly==10.0
HyperPyYAML==1.2.2
identify==2.6.12
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.8.1
k2==0.3.0
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.2
lightning-utilities==0.14.3
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nodeenv==1.9.1
noisereduce==3.0.3
numba==0.61.2
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.22.0
optuna==4.4.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
pre_commit==4.2.0
primePy==1.3
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycodestyle==2.11.0
pycparser==2.22
pydoclint==0.4.1
pyflakes==3.2.0
Pygments==2.19.2
pyparsing==3.2.3
pyreadline3==3.5.4
pytest==7.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic-bin==0.4.14
pytorch-lightning==2.5.2
pytorch-metric-learning==2.8.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
SQLAlchemy==2.0.41
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1+cu118
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.1
torchmetrics==1.7.4
torchvision==0.22.1+cu118
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
wcwidth==0.2.13
yamllint==1.35.1
yarl==1.20.1
